requirements.txt
colorama
defusedxml
kaitaistruct
pycryptodome